+++ zk_full/jni/logic/testSliderLogic.cc
@@ -77,7 +77,7 @@ static void onProgressChanged_SeekBar2(ZKSeekBar *pSeekBar, int progress) {
 	printf("seekbar2 pos:%d\n", progress);
 	mTextValuePtr->setText(progress);
 	BRIGHTNESSHELPER->setBrightness(progress);
-    BYTE data[2] = {0x00, progress};
+    BYTE data[2] = {0x00, (BYTE)progress};
     sendProtocol(CMD_SEEKBAR_LIGHT, data, sizeof(data));
 }
 
@@ -89,7 +89,7 @@ static void onProgressChanged_SeekBar1(ZKSeekBar *pSeekBar, int progress) {
 //	mCirclebar1Ptr->setProgress(progress);
 	printf("seekbar1 pos:%d\n", progress);
 	mTextview3Ptr->setText(progress);
-	BYTE data[2] = {0x00, progress};
+	BYTE data[2] = {0x00, (BYTE)progress};
 	sendProtocol(CMD_SEEKBAR_TEMPERATURE, data, sizeof(data));
 	printf("seekbar1 pos xx:%d\n", progress);
 }
